+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+1.0.2 (12/12/2017)
+
+- Fixed repeated commands
+- Windows version now shown
+- Username of client now shown
+- Antivirus shown 
+
 1.0.1 (12/12/2017)
 
 - Fixed incorrect CHANGELOG.md